+++ runelite-client/src/main/java/net/runelite/client/plugins/customizesidebar/CustomizeSidebarConfig.java
@@ -22,17 +22,17 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.client.plugins.reordersidebar;
+package net.runelite.client.plugins.customizesidebar;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.Keybind;
 
-@ConfigGroup(ReorderSidebarConfig.CONFIG_GROUP)
-public interface ReorderSidebarConfig extends Config
+@ConfigGroup(CustomizeSidebarConfig.CONFIG_GROUP)
+public interface CustomizeSidebarConfig extends Config
 {
-	String CONFIG_GROUP = "reorderSidebar";
+	String CONFIG_GROUP = "customizeSidebar";
 
 	@ConfigItem(
 		keyName = "dragRequiresHotkey",
@@ -56,3 +56,4 @@ public interface ReorderSidebarConfig extends Config
 		return Keybind.SHIFT;
 	}
 }
+
